- Updated to remove singleton of WorkoutSessionViewModel - WorkoutManager is now the single source of truth for the persistence of in progress workout data - Sets now are added properly to WorkoutSessionScreen

+++ shared/src/commonMain/kotlin/com/squalec/liftingtracker/appdatabase/models/UserWorkoutSession.kt
@@ -12,6 +12,6 @@ data class  UserWorkoutSession(
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "id") val id: Int = 0,
     @ColumnInfo(name = "date") val date: String = "",
-    @ColumnInfo(name = "duration") val duration: Int = 0,
+    @ColumnInfo(name = "duration") val duration: Long = 0,
     @ColumnInfo(name = "calories_burned") val caloriesBurned: Int = 0,
 )
